requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
datasets==4.6.1
debugpy==1.8.19
decorator==5.2.1
dill==0.4.0
executing==2.2.1
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
gitpython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.8.0
jupyter-core==5.9.1
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
pandas==3.0.0
parso==0.8.5
peft==0.18.1
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyyaml==6.0.3
pyzmq==27.1.0
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
sentry-sdk==2.50.0
setuptools==80.10.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
triton==3.6.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
wandb==0.24.0
wcwidth==0.3.3
xxhash==3.6.0
yarl==1.22.0
